requirements.txt
Flask==3.0.2
psutil==5.9.8
pefile==2023.2.7
puremagic==1.21
oletools==0.60.1
yara-python==4.5.0
werkzeug==3.0.1
requests==2.31.0
python-nmap==0.7.1
pysnmp-lextudio==5.0.34
pyasn1==0.4.8
pycryptodomex==3.18.0
dnspython==2.6.1
whois
beautifulsoup4
